fix(FFS-1183): cast geo_latitude/longitude to numeric in MIG_3057/3058
Production apply (2026-04-07) of MIG_3058 failed with:

  ERROR: function sot.service_area_membership(double precision, double precision) does not exist

because ops.intake_submissions.geo_latitude and .geo_longitude are
stored as double precision, but sot.service_area_membership takes
NUMERIC parameters. Postgres does not auto-cast between numeric types
in function resolution, even when the values would fit.

The trigger function ops.compute_service_area_status() has the same
bug — it would break the first time a real intake submission sets a
geo_latitude or the existing trigger fires on any UPDATE. Caught in
prod because all 1303 existing submissions have NULL coords, so the
trigger has never actually fired against real data yet.

Both call sites now use explicit ::numeric casts. The live trigger
function in prod was patched directly via psql during the apply;
this commit syncs the fix back to the source file so fresh installs
get the correct version.

Verified after applying the patch: updating a submission's
geo_latitude to (38.4404, -122.7141) correctly returns 'in' via the
BEFORE UPDATE trigger.

+++ sql/schema/v2/MIG_3057__service_area_boundary.sql
@@ -191,9 +191,13 @@ BEGIN
   END IF;
 
   IF NEW.geo_latitude IS NOT NULL AND NEW.geo_longitude IS NOT NULL THEN
+    -- Cast to NUMERIC — geo_latitude/longitude are double precision,
+    -- sot.service_area_membership() takes NUMERIC. Explicit cast avoids
+    -- "No function matches" errors on fresh installs (verified 2026-04-07
+    -- during prod apply).
     NEW.service_area_status := sot.service_area_membership(
-      NEW.geo_latitude,
-      NEW.geo_longitude
+      NEW.geo_latitude::numeric,
+      NEW.geo_longitude::numeric
     );
     NEW.service_area_status_set_at := NOW();
     NEW.service_area_status_source := 'auto';
